+++ tf/modules/ecs/task_definition/variables.tf
@@ -49,6 +49,25 @@ variable "efs_volumes" {
   default = []
 }
 
+variable "ebs_volumes" {
+  type = list(object({
+    name = string
+    ebs_volume_configuration = object({
+      encrypted        = optional(bool, true)
+      iops             = optional(number)
+      size_in_gib      = number
+      snapshot_id      = optional(string)
+      throughput       = optional(number)
+      volume_type      = optional(string, "gp3")
+      kms_key_id       = optional(string)
+      file_system_type = optional(string)
+      role_arn         = string
+    })
+  }))
+  default     = []
+  description = "List of EBS volumes to attach to the task"
+}
+
 variable "placement_constraints" {
   type = list(object({
     type       = string